requirement.txt
python==3.8
pytotch==1.11.0_cu113
torchvision==0.12.0+cu113
opencv-python==4.5
tensorboard
monai
medpy
SimpleITK
nibabel
einops